/tasks route done

+++ chunkmydocs/src/models/server/tasks.rs
@@ -3,8 +3,8 @@ use utoipa::ToSchema;
 
 #[derive(Debug, Serialize, Deserialize, Clone, ToSchema)]
 pub struct TasksQuery {
-    pub page: Option<i32>,
-    pub limit: Option<i32>,
+    pub page: Option<i64>,
+    pub limit: Option<i64>,
 }
